% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis.R
\name{kobo_analysis}
\alias{kobo_analysis}
\title{Generate the analysis}
\usage{
kobo_analysis()
}
\description{
Automatically generate bar charts for each select_one and select_multiple for in the dataset. ggplot2 is used.
}
\examples{
kobo_analysis()

\dontrun{
kobo_analysis()
}


}
\author{
Elliott Messeiller
}
