% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis_plan.R
\name{kobo_analysis_plan}
\alias{kobo_analysis_plan}
\title{Import the data analysis plan from the XLS form}
\usage{
kobo_analysis_plan()
}
\value{
A file with all elements to get your data & form.
}
\description{
Add additional variables based on the data analysis plan to the data frame
}
\examples{
kobo_analysis_plan()

}
\author{
Elliott Messeiller
}
