% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_arrange_variablename.R
\name{kobo_arrange_variablename}
\alias{kobo_arrange_variablename}
\title{Replace / or : in variable name in order to use the dictionnary}
\usage{
kobo_arrange_variablename(data)
}
\arguments{
\item{data}{dataframe with Variables to be renamed}
}
\description{
The character to be replaced - coudl be a "/" or a ":"
}
\examples{
kobo_arrange_variablename()

\dontrun{
kobo_arrange_variablename(data)
}


}
\author{
Edouard Legoupil
}
