% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_one.R
\name{kobo_bar_one}
\alias{kobo_bar_one}
\title{Generate bar Chart - frequency - for select_one questions}
\usage{
kobo_bar_one(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\description{
Automatically generate bar chart for each of the select_one question in the dataset. ggplot2 is used.
}
\examples{
kobo_bar_one()

\dontrun{
kobo_bar_one()
}


}
\author{
Edouard Legoupil, Elliott Messeiller
}
