% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_one_facet.R
\name{kobo_bar_one_facet}
\alias{kobo_bar_one_facet}
\title{Generate faceted frequency bar chart}
\usage{
kobo_bar_one_facet(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for proper shiny app path}
}
\description{
Automatically generate faceted chart for select one variable.. ggplot2 is used.
}
\examples{
kobo_bar_one_facet()

\dontrun{
kobo_bar_one_facet()
}


}
\author{
Edouard Legoupil, Elliott Messeiller
}
