% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_boxplot_facet.R
\name{kobo_boxplot_facet}
\alias{kobo_boxplot_facet}
\title{Generate histogramm plots based on dates}
\usage{
kobo_boxplot_facet()
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate boxplot. ggplot2 is used.
}
\examples{
kobo_boxplot_facet()

\dontrun{
kobo_boxplot_facet()
}


}
\author{
Edouard Legoupil, Elliott Messeiller
}
