% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_consolidateone.R
\name{kobo_consolidateone}
\alias{kobo_consolidateone}
\title{Merge disagregated select_one variable}
\usage{
kobo_consolidateone(data, dico)
}
\arguments{
\item{data}{original dataset}

\item{dico}{dictionnary}
}
\value{
A "data.table" with additional select_one variable.

data
}
\description{
Merge disagregated select_one variable
}
\examples{
kobo_consolidateone()

\dontrun{
kobo_consolidateone("myform.xls")
}

}
\author{
Edouard Legoupil
}
