% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_correlation.R
\name{kobo_correlation}
\alias{kobo_correlation}
\title{Generate histogramm plots based on dates}
\usage{
kobo_correlation()
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate maps for all nominal & ordinal variables based on dates. ggplot2 is used.
}
\examples{
kobo_correlation()

\dontrun{
kobo_correlation(S)
}


}
\author{
Edouard Legoupil
}
