% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_datasets2.R
\name{kobo_datasets2}
\alias{kobo_datasets2}
\title{Dataset list}
\usage{
kobo_datasets2(user, api)
}
\arguments{
\item{user}{Optional. A single string indicating the username and password
(in the form of \code{"username:password"}), or a character vector or list,
length 2, with the first value being the "username", and the second being
the "password".}

\item{api}{The URL at which the API can be accessed. Defaults to "kobo",
which loads the KoBo Toolbox API.}
}
\value{
A data.table containing details about the datasets available,
including items like the "title", "id", and "submission".
}
\description{
Lists the Datasets Available including count of submission.
- description    id  -    id_string - title   -   url
}
\examples{
kobo_datasets2()

}
\author{
Edouard Legoupil
}
