% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_dico.R
\name{kobo_dico}
\alias{kobo_dico}
\title{Data dictionnary}
\usage{
kobo_dico(mainDir = "")
}
\arguments{
\item{mainDir}{}
}
\value{
A "data.table" with the full data dictionnary. To be used in the rest of the analysis.
}
\description{
Produce a data dictionnary based on the xlsform for the project
}
\examples{
kobo_dico()

\dontrun{
kobo_dico()
}

}
\author{
Edouard Legoupil, Elliott Messeiller
}
