% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_encode.R
\name{kobo_encode}
\alias{kobo_encode}
\title{Encode variable}
\usage{
kobo_encode(data, dico)
}
\arguments{
\item{data}{Dataframe to relabel}

\item{dico}{Data dictionnary generated from kobo_dico}
}
\value{
A "data.table" with the full data.label. To be used for graphs generation.
}
\description{
Insert the full label in data frame based on dictionnary
}
\examples{
kobo_encode()

\dontrun{
kobo_encode(data, dico)
}

}
\author{
Edouard Legoupil
}
