% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_encode_repeat.R
\name{kobo_encode_repeat}
\alias{kobo_encode_repeat}
\title{Encode variable}
\usage{
kobo_encode_repeat(data, dico)
}
\arguments{
\item{data}{Dataframe to re-label}

\item{dico}{Data dictionnary generated from kobo_dico}
}
\value{
A "data.table" with the full data.label. To be used for graphs generation.
}
\description{
Insert the full label in data frame based on dictionnary - used when data is exported through briefcase because of repeated element in the dataset. In the this case, merge is done on name instead of fullname.
}
\examples{
kobo_encode_repeat()

\dontrun{
kobo_encode_repeat(data, dico)
}

}
\author{
Edouard Legoupil
}
