% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_forminfo.R
\name{kobo_forminfo}
\alias{kobo_forminfo}
\title{Get form attributes}
\usage{
kobo_forminfo(formid, user = NULL, api = api)
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string).}

\item{user}{Optional. A single string indicating the username and password
(in the form of \code{"username:password"}), or a character vector or list,
length 2, with the first value being the "username", and the second being
the "password".}

\item{api}{The URL at which the API can be accessed.
Defaults to "unhcr", which loads the UNHCR KoBo Toolbox API.}
}
\value{
A "data.table" with the full forminfo.
The forminfo would be named in the form of "data_formid".

The URL of the form based on form id.
}
\description{
Obtain form info in order to correctly retrieve the form.
}
\examples{
kobo_forminfo()
#' @examples
\dontrun{
kobo_forminfo("15051")
kobo_forminfo("31511", api = "unhcr")
}


}
\author{
Edouard Legoupil
}
