% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_histo.R
\name{kobo_histo}
\alias{kobo_histo}
\title{Generate histograme for all integer questions}
\usage{
kobo_histo()
}
\arguments{
\item{data}{kobodatset to use}

\item{dico}{( generated from kobo_dico)}
}
\description{
Automatically generate histogrammes for each of the integer questions in the dataset. ggplot2 is used.
}
\examples{
kobo_histo()

\dontrun{
kobo_histo()
}


}
\author{
Edouard Legoupil
}
