% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_label.R
\name{kobo_label}
\alias{kobo_label}
\title{Label Variable}
\usage{
kobo_label(datalabel, dico)
}
\arguments{
\item{dico}{( generated from kobo_dico)}

\item{data}{.}
}
\value{
A "data.table" with the full data.label. To be used for graphs generation.
}
\description{
Insert the full label in data frame based on dictionnary
}
\examples{
kobo_label()

\dontrun{
kobo_label(data, dico)
}

}
\author{
Edouard Legoupil
}
