% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_projectconfig.R
\name{kobo_projectconfig}
\alias{kobo_projectconfig}
\title{Project configuration file}
\usage{
kobo_projectconfig()
}
\value{
A file with all elements to get your data & form.
}
\description{
Write all necessary configuration files for your project
}
\examples{
kobo_projectconfig()

}
\author{
Edouard Legoupil
}
