% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_projectinit.R
\name{kobo_projectinit}
\alias{kobo_projectinit}
\title{Analysis project initiation}
\usage{
kobo_projectinit()
}
\value{
A structure of directory and scripts in order to set up quickly a project.
}
\description{
Create analysis project structure
}
\examples{
kobo_projectinit()

}
\author{
Edouard Legoupil, Elliott MEesseiller
}
