% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_question.R
\name{kobo_question}
\alias{kobo_question}
\title{Generates graphics and basic information based on the type of question and if there's a disaggregation}
\usage{
kobo_question(question, mainDir = "")
}
\arguments{
\item{question}{Name of the question to be treated, as in "fullname" column in dico}

\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\description{
Automatically generates bar charts, histograms, and basic information of the question based on the question type to generate the report.
}
\examples{
kobo_question("s2.beneficiary_code")

\dontrun{
kobo_question("s2.beneficiary_code")
}


}
\author{
Elliott Messeiller
}
