% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_report.R
\name{kobo_report}
\alias{kobo_report}
\title{Generate the analysis}
\usage{
kobo_report(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for proper shiny app path}
}
\description{
Automatically generate the report with all variable analysed
}
\examples{
kobo_report()

\dontrun{
kobo_report()
}


}
\author{
Edouard Legoupil, Elliott Messeiller
}
