% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_surveyname.R
\name{kobo_surveyname}
\alias{kobo_surveyname}
\title{Extract Survey name from XlsForm}
\usage{
kobo_surveyname(form)
}
\arguments{
\item{form}{}
}
\description{
parse xlsfrom
}
\examples{
kobo_surveyname()

\dontrun{r
kobo_surveyname(form)
}


}
\author{
Edouard Legoupil
}
