% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_text_cloud.R
\name{kobo_text_cloud}
\alias{kobo_text_cloud}
\title{text Could}
\usage{
kobo_text_cloud(data, dico)
}
\arguments{
\item{form}{The full filename of the form to be accessed (xls or xlsx file).
It is assumed that the form is stored in the data folder.}
}
\value{
A "data.table" with the full data dictionnary. To be used in the rest of the analysis.
}
\description{
Produce word cloud visualisation for the text questions. Can be also effective to see the results of or_other questions.
}
\examples{
kobo_text_cloud()

\dontrun{
kobo_text_cloud("myform.xls")
}

}
\author{
Edouard Legoupil
}
