% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_trend.R
\name{kobo_trend}
\alias{kobo_trend}
\title{Generate histogramm plots based on dates}
\usage{
kobo_trend(data, date, dico)
}
\arguments{
\item{data}{kobodatset to use}

\item{date}{field of date type used to generare trends}

\item{dico}{( generated from kobo_dico)}

\item{duration}{number of days in the past}
}
\description{
Automatically generate histogramm for all nominal & ordinal variables based on dates. ggplot2 is used.
}
\examples{
kobo_bar_trend()

\dontrun{
kobo_trend(data, date, dico)
}


}
\author{
Edouard Legoupil
}
