% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_bar_multi_facet.R
\name{kobo_bar_multi_facet}
\alias{kobo_bar_multi_facet}
\title{Generate frequency bar chart for select_multiple variable}
\usage{
kobo_bar_multi_facet(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for proper shiny app path}
}
\description{
Automatically generate faceted chart for select multiple variables. ggplot2 is used.
}
\examples{
kobo_bar_multi_facet()

\dontrun{
kobo_bar_multi_facet()
}


}
\author{
Edouard Legoupil
}
