% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_histo.R
\name{kobo_histo}
\alias{kobo_histo}
\title{Generate histograme for all integer questions}
\usage{
kobo_histo(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for proper shiny app path}
}
\description{
Automatically generate histogrammes for each of the integer questions in the dataset. ggplot2 is used.
}
\examples{
kobo_histo()

\dontrun{
kobo_histo()
}


}
\author{
Edouard Legoupil
}
