% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_weight.R
\name{kobo_weight}
\alias{kobo_weight}
\title{Weight the data}
\usage{
kobo_weight(mainDir = "")
}
\arguments{
\item{mainDir}{Path to the project's working directory: mainly for shiny app}
}
\description{
Automatically weight the data according to the information of 0-config.R
}
\examples{
kobo_weight()

\dontrun{
kobo_weight()
}


}
\author{
Elliott Messeiller
}
